# src/services/rest-api/swagger-json.json
{"openapi":"3.0.0","info":{"title":"Mossem API","description":"REST API for Mossem","version":"1.0","contact":{}},"tags":[{"name":"food mossem","description":""}],"servers":[{"url":"/mossem-dfa21/asia-northeast1/restapi"},{"url":"https://asia-northeast1-mossem-dfa21.cloudfunctions.net/restapi"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"DanceDto":{"type":"object","properties":{"id":{"type":"string"},"url":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}},"required":["id","url","name","description"]},"InstrumentDto":{"type":"object","properties":{"id":{"type":"string"},"url":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"dateaquired":{"type":"string"},"quantity":{"type":"string"},"status":{"type":"string"}},"required":["id","url","name","description","dateaquired","quantity","status"]},"SongDto":{"type":"object","properties":{"id":{"type":"string"},"url":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"songwriter":{"type":"string"},"performedplaces":{"type":"string"},"datecreated":{"type":"string"}},"required":["id","url","name","description","songwriter","performedplaces","datecreated"]},"PictureDto":{"type":"object","properties":{"id":{"type":"string"},"url":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}},"required":["id","url","name","description"]},"EquipmentDto":{"type":"object","properties":{"id":{"type":"string"},"url":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"dateaquired":{"type":"string"},"quantity":{"type":"string"},"status":{"type":"string"}},"required":["id","url","name","description","dateaquired","quantity","status"]},"RegisterUserDto":{"type":"object","properties":{"fullName":{"type":"string","default":"Muhammad Ibrahim Lucman"},"email":{"type":"string","default":"admin@itsmarawi.net"},"password":{"type":"string","default":"oddminyousir"}},"required":["fullName","email","password"]},"AccessTokenDto":{"type":"object","properties":{"accessToken":{"type":"string","minLength":5},"refreshToken":{"type":"string","minLength":5}},"required":["accessToken"]},"RefreshDto":{"type":"object","properties":{"refresh_token":{"type":"string","minLength":5}},"required":["refresh_token"]},"UserDto":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"},"userType":{"type":"string"}},"required":["fullName","email","userType"]}}},"paths":{"/":{"get":{"operationId":"about","summary":"About this REST API","parameters":[],"responses":{"200":{"description":""}}}},"/dance":{"post":{"operationId":"AddDance","summary":"Add new dance","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DanceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DanceDto"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"GetDances","summary":"Get information of all dance","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DanceDto"}}}}}}}},"/dance/{id}":{"get":{"operationId":"GetDance","summary":"Get information of dance with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DanceDto"}}}}}},"delete":{"operationId":"DeleteDance","summary":"Delete dance with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DanceDto"}}}}},"security":[{"bearer":[]}]},"put":{"operationId":"UpdateDance","summary":"update dance","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DanceDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/instrument":{"post":{"operationId":"AddInstrument","summary":"Add new instrument","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentDto"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"GetInstruments","summary":"Get information of all instrument","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InstrumentDto"}}}}}}}},"/instrument/{id}":{"get":{"operationId":"GetInstrument","summary":"Get information of instrument with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentDto"}}}}}},"delete":{"operationId":"DeleteInstrument","summary":"Delete instrument with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentDto"}}}}},"security":[{"bearer":[]}]},"put":{"operationId":"UpdateInstrument","summary":"update instrument","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstrumentDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/song":{"post":{"operationId":"AddSong","summary":"Add new song","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SongDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SongDto"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"GetSongs","summary":"Get information of all song","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SongDto"}}}}}}}},"/song/{id}":{"get":{"operationId":"GetSong","summary":"Get information of song with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SongDto"}}}}}},"delete":{"operationId":"DeleteSong","summary":"Delete song with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SongDto"}}}}},"security":[{"bearer":[]}]},"put":{"operationId":"UpdateSong","summary":"update song","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SongDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/picture":{"post":{"operationId":"AddPicture","summary":"Add new picture","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PictureDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PictureDto"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"GetPictures","summary":"Get information of all picture","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PictureDto"}}}}}}}},"/picture/{id}":{"get":{"operationId":"GetPicture","summary":"Get information of picture with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PictureDto"}}}}}},"delete":{"operationId":"DeletePicture","summary":"Delete picture with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PictureDto"}}}}},"security":[{"bearer":[]}]},"put":{"operationId":"UpdatePicture","summary":"update picture","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PictureDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/equipment":{"post":{"operationId":"AddEquipment","summary":"Add new equipment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentDto"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"GetEquipments","summary":"Get information of all equipment","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EquipmentDto"}}}}}}}},"/equipment/{id}":{"get":{"operationId":"GetEquipment","summary":"Get information of equipment with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentDto"}}}}}},"delete":{"operationId":"DeleteEquipment","summary":"Delete equipment with id {id}","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentDto"}}}}},"security":[{"bearer":[]}]},"put":{"operationId":"UpdateEquipment","summary":"update equipment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipmentDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"register","summary":"Register new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"201":{"description":""}}}},"/auth/login":{"post":{"operationId":"login","summary":"Login User","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDto"}}}}}}},"/auth/refresh_token":{"post":{"operationId":"refreshToken","summary":"Refresh Token","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDto"}}}}}}},"/auth/logout":{"post":{"operationId":"logout","summary":"logout given user","parameters":[],"responses":{"201":{"description":""}}}},"/auth/profile":{"get":{"operationId":"getProfile","summary":"get profile info","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearer":[]}]}}}}